moves MML parsing related functions to 3d-web-avatar package
