+++ example/assets/avatar/parts/SK_Andor_Body_02.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2aea0b3041cc9f60462ba04749820dff667b87ddd3ff37dd884edb0903e59249
+size 12095516
