+++ example/assets/avatar/parts/SK_Andor_Body_02_2.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:18caa586fc971153a20c6795eb3c07430f5a9c6668b36547d216d77f1dc3f431
+size 11517512
