+++ example/assets/avatar/parts/SK_Andor_FullBody_01.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:102dfd271dc08b98b9b0f0cf7f75bb034095e67b209112bad95ded0aef12ebea
+size 2408616
