+++ example/assets/avatar/parts/SK_Andor_FullBody_03.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2317e9125540a33cd9f3840ab5f858241dcbc7cc1cc9ae25a4fc966b5615c0fa
+size 1003800
