+++ example/assets/avatar/parts/test2.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:098b29e87bbb8ca2b6749bde97295891e6c195e52e44f2b2d7783b69c1c01c65
+size 18976732
