+++ example/assets/avatar/parts/test3.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e1e51b119775311e686e7f70fd7e65896ca9347c73113296a5d2815e0ddbe88a
+size 3535512
