+++ example/assets/avatar/parts/test5.glb
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ae8110d45a094ca3993d41e8fdfa13753fffd471c986579ef3beaa65c67e9d6f
+size 6545464
