+++ packages/3d-web-avatar/src/helpers/parseMMLDescription.ts
@@ -1,4 +1,11 @@
-import { MMLCharacterDescription, MMLCharacterDescriptionPart } from "./AvatarEditor";
+export type MMLCharacterDescriptionPart = {
+  url: string;
+};
+
+export type MMLCharacterDescription = {
+  base: MMLCharacterDescriptionPart;
+  parts: MMLCharacterDescriptionPart[];
+};
 
 export type LoadingErrors = string[];
