+++ packages/3d-web-avatar/src/index.ts
@@ -1,3 +1,10 @@
 export { Character } from "./character/Character";
 export { TimeManagerInterface } from "./character/types";
 export { ModelLoader } from "./character/ModelLoader";
+export { findAssetsInCollection } from "./helpers/findAssetsInCollection";
+export {
+  type LoadingErrors,
+  type MMLCharacterDescriptionPart,
+  type MMLCharacterDescription,
+  parseMMLDescription,
+} from "./helpers/parseMMLDescription";
